アドレスの型をintからuint16に変更

+++ bus/ppubus.go
@@ -8,10 +8,10 @@ func NewPpuBus(characterRam *Ram) PpuBus {
 		return PpuBus{CharacterRam: *characterRam}
 }
 
-func (ram PpuBus) readByPpu(addr int) byte {
+func (ram PpuBus) readByPpu(addr uint16) byte {
 	return ram.CharacterRam.Read(addr)
 }
 
-func (ram *PpuBus) writeByPpu(addr int, val byte) {
+func (ram *PpuBus) writeByPpu(addr uint16, val byte) {
 	ram.CharacterRam.Write(addr, val)
 }
