+++ cpubus/cpubus.go
@@ -15,7 +15,7 @@ type CpuBus struct {
 }
 
 
-func (this CpuBus) ReadWord(addr int) uint16 {
+func (this CpuBus) ReadWord(addr uint16) uint16 {
 	/**
 	fn read_word(&mut self, addr: u16) -> u16 {
 		let lower = self.read(addr) as u16;
@@ -23,9 +23,9 @@ func (this CpuBus) ReadWord(addr int) uint16 {
 		(upper << 8 | lower) as u16
 	}
 	 */
-	lower := uint16(this.ReadByCpu(addr))
-	upper := uint16(this.ReadByCpu(addr + 1))
-	return upper << 8 | lower
+	lower := this.ReadByCpu(addr)
+	upper := this.ReadByCpu(addr + 1)
+	return uint16(upper << 8 | lower)
 }
 
 
@@ -40,7 +40,7 @@ func NewCpuBus(ram bus.Ram, programRom bus.Rom, ppu ppu.Ppu, dma dma.Dma) CpuBus
 	return cpuBus
 }
 
-func (this CpuBus) ReadByCpu(addr int) byte {
+func (this CpuBus) ReadByCpu(addr uint16) byte {
 	switch {
 	case 0x0000 <= addr && addr <= 0x1FFF:
 		return this.Ram.Read(addr)
@@ -71,7 +71,7 @@ func (this CpuBus) ReadByCpu(addr int) byte {
 }
 
 
-func (this *CpuBus) WriteByCpu(addr int, data byte) {
+func (this *CpuBus) WriteByCpu(addr uint16, data byte) {
 	switch {
 	case 0x0000 <= addr && addr <= 0x1FFF:
 		this.Ram.Write(addr&0x07FF, data)
